% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIS.R
\docType{data}
\name{AIS}
\alias{AIS}
\title{Australian institute of sport data}
\format{
A matrix with 202 observations on the following variables:
\describe{
\item{Sex}{0 = Male or 1 = Female.}
\item{Ht}{Height (in cm).}
\item{LBM}{Lean body mass (in Kg).}
\item{RCC}{Red cell count.}
\item{Hc}{Hematocrit.}
\item{Hg}{Hemoglobin.}
\item{SSF}{Sum of skin folds.}
\item{Bfat}{Body fat percentage.}
}
}
\source{
This dataset is a subset of the \code{ais} dataset contained in the \code{alr4} R package.
}
\usage{
data(AIS)
}
\description{
A dataset containing biometrical measurements for two categories of athletes collected at the Australian Institute of Sport.
}
\references{
Weisberg Sanford (2018). alr4: Data to Accompany Applied Linear Regression 4th Edition.
\url{https://CRAN.R-project.org/package=alr4}.
}
\keyword{datasets}
